Samples/policy-generator-app/requirements.txt
streamlit>=1.30.0
openai>=1.10.0
python-dotenv>=1.0.0
pypdf>=4.0.0
